backport dexterity copy flag patch.
**Problem:**

When copying a DX container which has AT children, the UID of the AT
children was not updated.
The reason for the error is that the DX container copy did not have the
`_v_is_cp` flag while the AT children were processed and thus the flag
was not properly delegated.

**Solution:**

By copying the `_v_is_cp` and `_v_cp_refs flags` to the copy we have the
same behavior as it used to be with AT, which does fix the error.

- Issue: https://github.com/plone/Products.CMFPlone/issues/1735
- Plone 4 fix: https://github.com/plone/plone.dexterity/pull/60
- Plone 5 fix: https://github.com/plone/plone.dexterity/pull/61