Fix edge case with collective.indexing and other 3rd party packages and rename a nested structure.
The queue needs to be processed, since the `renameObjectsByPaths`
script allows the user to rename the object and also sets a new
title if he wants.
In some circumstances this leads to a inconsistent catalog state.
Mainly if other event handler also triggers a `process` queue by
asking the catalog for something.
The result was for example a "reindex" of a already deleteobject.